Use the full Rebin signature
Otherwise the vert and lat error bands give ambiguous overload warnings.
We want to support the full behavior in the long run anyway.

The additional arguments that allow for a clone to be produced in rebin
are not supported.  If code tries to use them you get an exception.

small fixes to makefile: remove -Wl and fix output redirect for lib
directory test