patches to env_set
define ROOMU_SYS by audodetecting the root directory in the env_set
script if possible.  if autodetect is not possible, print error.

prepend ROOTSYS/lib to ld_lib_path instead of setting (looked like
typo).  if ROOTSYS is defined then can we assume that paths are also set
for root?